pyproject.toml
[project]
name = "nonebot-plugin-neuro-draw"
version = "0.1.0"
description = "✨ Neuro 每日一签的 NoneBot 插件 ✨"
authors = [
    {name = "zhaomaoniu", email = "2667292003@qq.com"},
]
dependencies = ["nonebot2>=2.4.1", "nonebot-plugin-localstore>=0.7.3", "nonebot-plugin-alconna>=0.54.2", "aiohttp>=3.11.11"]
requires-python = ">=3.9"
readme = "README.md"
license = {text = "MIT"}

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"


[tool.pdm]
distribution = true
